Moved and renamed some stuff to keep it separate from legacy stuff.
