+++ cache/legacy.go
@@ -2,7 +2,6 @@ package cache
 
 import (
 	"github.com/prebid/prebid-server/pbs/buckets"
-	"encoding/json"
 )
 
 type Domain struct {
@@ -36,18 +35,4 @@ type AccountsService interface {
 type ConfigService interface {
 	Get(string) (string, error)
 	Set(string, string) error
-}
-
-// ConfigFetcher knows how to fetch OpenRTB configs by id.
-// Implementations must be safe for concurrent access by multiple goroutines.
-//
-// A config is basically a "partial" OpenRTB request.
-// The Endpoint merges these into the HTTP Request JSON before unmarhsalling it
-// into the OpenRTB Request which gets sent into the Exchange.
-type ConfigFetcher interface {
-	// GetConfigs fetches configs for the given IDs.
-	// The returned map will have keys for every ID, unless errors exist.
-	//
-	// The returned objects should only be read from--never written to.
-	GetConfigs(ids []string) (map[string]json.RawMessage, []error)
 }
